Fix: Palette Loader (External Palette) not applying correct color count is fixed.
