+++ src/ImageEditor/PaletteImporter/PaletteLoader.cs.uid
@@ -1,1 +1,1 @@
-uid://0m4e2jg22pby
+uid://d0xqdcorc7y2b
