--- src/ImageEditor/PaletteImporter/PaletteLoaderFlowUi.cs.uid
@@ -1,1 +1,0 @@
-uid://d0xqdcorc7y2b
